Allow reset by user choice

+++ src/main/java/edu/csumb/spring19/capstone/controllers/CommonDataController.java
@@ -8,6 +8,7 @@ import edu.csumb.spring19.capstone.models.common.CommonData;
 import edu.csumb.spring19.capstone.models.common.CommonInitService;
 import edu.csumb.spring19.capstone.repos.CommonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
@@ -15,7 +16,11 @@ import java.util.Optional;
 @RestController
 @CrossOrigin("*")
 public class CommonDataController{
+    @Value("${ALLOW_COMMON_RESET:false}")
+    private Boolean allowReset;
 
+    @Autowired
+    private CommonInitService commonInitService;
     @Autowired
     private CommonRepository commonRepository;
 
@@ -43,9 +48,11 @@ public class CommonDataController{
 
     @DeleteMapping({"/admin/reset"})
     public RestDTO resetCommonData() {
-        commonRepository.deleteAll();
-        CommonInitService.initDatabase();
-        return new RestSuccess();
+        if (allowReset) {
+            commonRepository.deleteAll();
+            commonInitService.initDatabase();
+            return new RestSuccess();
+        } else return new RestFailure("This functionality is disabled.");
     }
 }
